requirements.txt
fastapi
uvicorn
python-dotenv
openai
gspread
oauth2client
Pillow
rembg
python-telegram-bot==13.15
firebase-admin
reportlab
yookassa
slowapi
pydantic[email]
redis
